pyproject.toml
[project]
name = "ctbto"
version = "0.1.0"
dependencies = [
    "pycryptodome",
    "matplotlib",
    "numpy",
    "tqdm",
    "SciencePlots",
    "black"
]

[project.scripts]
ctbto-benchmark = "ctbto.cli.main:main"
